requirements.txt
tqdm>=4.65.0
customtkinter>=5.2.0
Pillow>=9.0.0
